requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
colorama==0.4.4
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
flake8==4.0.1
iniconfig==1.1.1
mccabe==0.6.1
model-bakery==1.4.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==6.2.5
pytest-django==4.5.2
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
